refactor: move CareerCard and CareerCards file position to make structure more readable
