+++ components/templates/Career/CareerCard/CareerCard.module.scss
