+++ components/templates/Career/CareerCard/index.ts
